Man, Play's Iteratee model is rough...

+++ app/maptower/map/Point.scala
@@ -4,8 +4,8 @@ import com.mongodb.BasicDBList
 import com.mongodb.casbah.commons.MongoDBList
 
 object Point {
-  val empty = new Point(Int.MinValue, Int.MinValue)
-  val center = new Point(0, 0)
+  val empty = new Point(Double.MinValue, Double.MinValue)
+  val center = new Point(0.0, 0.0)
 
   def apply(objList: BasicDBList): Point = {
     new Point(objList.get(0).asInstanceOf[Double], objList.get(1).asInstanceOf[Double])
